Move Web.Services specific code to that project
