+++ sipsorcery-core/SIPSorcery.Sys/SIPSorcery.Sys.csproj
@@ -131,8 +131,6 @@
     <Compile Include="AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Auth\ServiceAuthToken.cs" />
-    <Compile Include="Auth\SIPSorcerySecurityHeader.cs" />
     <Compile Include="AzureTraceAppender.cs" />
     <Compile Include="CRC32.cs" />
     <Compile Include="Crypto\PasswordHash.cs" />
