+++ sipsorcery-core/SIPSorcery.Web.Services/Auth/ServiceAuthToken.cs
